refactor: organizes BrowserSelect exports in their respective folders
