+++ src/components/admin/browser-select/index.ts
@@ -1,2 +1,2 @@
 export { default as BrowserSelect } from './BrowserSelect.vue';
-export type { BrowserSelectProps } from './type';
+export type { BrowserSelectProps } from './types';
